refactor: LocalizedLiteral / String Extension 파일 분리
- ViewController에 포함되어있던 LocalizedLiteral과 String Extension 파일을 분리했습니다.
