+++ LoginProfile/LocalizedLiteral.swift
@@ -1,0 +1,18 @@
+//
+//  LocalizedLiteral.swift
+//  LoginProfile
+//
+//  Created by kakao on 2023/01/03.
+//
+
+import Foundation
+
+enum LocalizedLiteral {
+    static let welcome = "welcome_message"
+    static let info = "info_message"
+    static let idText = "id_text_field"
+    static let passwordText = "password_text_field"
+    static let signIn = "sign_in_button"
+    static let signUp = "sign_up_button"
+    static let findAccount = "find_account_button"
+}
