+++ LoginProfile/String+Extension.swift
@@ -1,0 +1,14 @@
+//
+//  String+Extension.swift
+//  LoginProfile
+//
+//  Created by kakao on 2023/01/03.
+//
+
+import Foundation
+
+extension String {
+    var localized: String {
+        return NSLocalizedString(self, comment: "")
+    }
+}
